requirements.txt
pyodbc>=4.0.39
python-dotenv==1.0.0
jinja2==3.1.2
markdown==3.5.1
azure-identity==1.15.0
colorama==0.4.6
click==8.1.7
schedule==1.2.0
requests>=2.31.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0

# Advanced Export Dependencies
xlsxwriter>=3.1.0
weasyprint>=59.0
reportlab>=4.0.0
python-docx>=0.8.11